requirements.txt
Pillow==3.2.0
python-resize-image==1.1.10
requests==2.9.1
